+++ CHANGELOG.md
@@ -7,6 +7,11 @@ This project adheres to [Semantic Versioning](http://semver.org).
 This document is formatted according to the principles of [Keep A CHANGELOG](http://keepachangelog.com).
 
 ## [Unreleased]
+### Changed
+- [Go] Upgraded messages to v22
+
+### Added
+- (i18n) Add Malayalam localization
 
 ## [26.2.0] - 2023-04-07
 ### Changed